refactor: index.js, App.js, barchart 폴더 분리
