+++ src/App.js
@@ -1,0 +1,11 @@
+import Board from './features/barchart/Board';
+
+function App() {
+    return (
+        <div className="App">
+            <Board />
+        </div>
+    );
+}
+
+export default App;
